+++ src/Request/Model/UpdateContact.php
@@ -8,4 +8,8 @@ class UpdateContact extends CreateContact
      * @var string
      */
     public $contact_id;
+    /**
+     * @var string
+     */
+    public $old_account_id;
 }